+++ h3ron/benches/collections_insert_get.rs
@@ -1,5 +1,4 @@
 use std::collections::BTreeMap;
-use std::iter::FromIterator;
 use std::time::Duration;
 
 use criterion::{criterion_group, criterion_main, Criterion};